+++ sdk/evaluation/azure-ai-evaluation/azure/ai/evaluation/_common/utils.py
@@ -270,3 +270,22 @@ def _validate_typed_dict(o: object, t: Type[T_TypedDict]) -> T_TypedDict:
         validate_annotation(v, annotations[k])
 
     return cast(T_TypedDict, o)
+
+def retrieve_content_type(messages, metric: str) -> str:
+    """Get the content type for service payload.
+
+    :param messages: The list of message to be annotation by evaluation service 
+    :type messages: list
+    :return: A text representing a content type. Example text, or images
+    :rtype: str
+    """
+    if metric.value == "protected_material":
+        return "image"
+    if len(messages) > 0:
+        for item in messages:
+            if "content" in item:
+                for content in item["content"]:
+                    if content.get("type") == "image_url":
+                        return "image"  
+        return "text"    
+    return "text"